refactor register profile
