+++ frontend/AngularApp/src/app/profile/profile.component.html
@@ -1,1 +1,7 @@
-<p>profile works!</p>
+<div class="app-wrapper">
+  <main class="main">
+    <app-card>
+      <h1 class="title">Profile Details</h1>
+    </app-card>
+  </main>
+</div>
