+++ frontend/AngularApp/src/app/profile/profile.component.ts
@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
+import { CardComponent } from '../UI/card/card.component';
 
 @Component({
   selector: 'app-profile',
   standalone: true,
-  imports: [],
+  imports: [CardComponent],
   templateUrl: './profile.component.html',
   styleUrl: './profile.component.scss',
 })
