--- frontend/AngularApp/src/app/register-profile/register-profile-form/register-profile-form.component.html
@@ -1,55 +1,0 @@
-<form class="form" #form="ngForm" (ngSubmit)="onSubmit(form)">
-    <h1 class="title">Profile Details</h1>
-    
-    <div class="form-group">
-      <label for="gender">Gender</label>
-      <select id="gender" name="gender" required ngModel>
-        <option value="" disabled selected>Select your gender</option>
-        <option value="male">Male</option>
-        <option value="female">Female</option>
-      </select>
-    </div>
-    
-    <div class="form-group">
-      <label for="preferences">Sexual Preferences</label>
-      <select id="preferences" name="preferences" required ngModel>
-        <option value="" disabled selected>Select your preference</option>
-        <option value="heterosexual">Heterosexual</option>
-        <option value="homosexual">Homosexual</option>
-        <option value="bisexual">Bisexual</option>
-      </select>
-    </div>
-    
-    <div class="form-group">
-      <label for="bio">Biography</label>
-      <textarea id="bio" name="bio" rows="4" required ngModel></textarea>
-    </div>
-    
-    <div class="form-group">
-        <label for="interests">Select Interests</label>
-        <div class="checkbox-group">
-          <div *ngFor="let tag of tagsList">
-            <input type="checkbox" id="{{ tag }}" name="interests" [value]="tag" (change)="onTagChange($event)" />
-            <label for="{{ tag }}">{{ tag }}</label>
-          </div>
-        </div>
-      </div>
-    
-    <div class="form-group">
-      <label for="pictures">Upload Pictures (max 5)</label>
-      <input type="file" id="pictures" name="pictures" accept="image/*" multiple required ngModel (change)="onFileChange($event)" />
-    </div>
-  
-    <div class="form-group">
-      <label for="profilePic">Select Profile Picture</label>
-      <select id="profilePic" name="profilePic" required ngModel>
-        <option *ngFor="let pic of uploadedPictures" [value]="pic">{{ pic }}</option>
-      </select>
-    </div>
-    
-    <button custom-button class="btn">Save Profile</button>
-    
-    <p>Already have an account?</p>
-    <a routerLink="/login">Login</a>
-  </form>
-
